fix foreach not defined on deploy

+++ backend/index.ts
@@ -165,8 +165,10 @@ HCServer.io.on("connect", (socket) => {
 
 
 fs.readdir("public/profile", (err, files) => {
-	files.forEach((file:any) => {
-		const filePath = path.join("public/profile", file);
-		fs.unlink(filePath, (err:any) => {});
-	});
+	if (files != undefined){
+		files.forEach((file:any) => {
+			const filePath = path.join("public/profile", file);
+			fs.unlink(filePath, (err:any) => {});
+		});
+	}
 });
